made tasks

+++ app/controllers/questions_controller.rb
@@ -5,7 +5,9 @@ class QuestionsController < ApplicationController
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @answer = question.answers.new
+  end
 
   def new; end
 
@@ -13,6 +15,7 @@ class QuestionsController < ApplicationController
 
   def create
     @question = Question.new(question_params)
+    @question.user = current_user
 
     if @question.save
       redirect_to @question, notice: 'Your question successfully created.'
@@ -30,7 +33,7 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    question.destroy if current_user.author?(question)
     redirect_to questions_path
   end
